chore: reorganize object detection utilities into package structure
